+++ CHANGELOG.md
@@ -3,6 +3,8 @@ docxBox Changelog
 
 V0.0.3 - Not finished yet
 -------------------------
+* Add Compress XMLs and zip (zpc)
+* Rename commands: rem -> rmt, zip -> zp
 * Add unzip and indent XML (uzi)
 * Improve list images (lsi): output file attributes